vsclib/freadall: cast save to size_t
MSVC complains

+++ vsclib/freadall.c
@@ -223,7 +223,7 @@ int vsc_freadalla_ex(void **ptr, size_t *size, FILE *f, VscFreadallInitProc init
 
         /* If we're already not at the start, don't allocate more than we need to. */
         if(state.file_size > 0) {
-            assert(save <= state.file_size);
+            assert((size_t)save <= state.file_size);
             state.file_size -= save;
         }
     }
